chore: add package init
