+++ scotus_dataset/__init__.py
@@ -1,0 +1,3 @@
+"""SCOTUS dataset utilities package."""
+
+__all__ = []
